enhance auth to rely on backend
