--- src/actions/login.ts
@@ -1,61 +1,0 @@
-"use server";
-
-import axios from "axios";
-import * as z from "zod";
-
-import { LoginSchema } from "~/schemas";
-
-const apiUrl = process.env.API_URL;
-
-export const loginUser = async (values: z.infer<typeof LoginSchema>) => {
-  const validatedFields = LoginSchema.safeParse(values);
-  if (!validatedFields.success) {
-    return {
-      error: "Login Failed. Please check your email and password.",
-    };
-  }
-  const { email, password } = validatedFields.data;
-  const payload = { email, password };
-  try {
-    const response = await axios.post(`${apiUrl}/api/v1/auth/login`, payload);
-
-    return {
-      status: response.status,
-      organisations: response.data.data.organisations,
-    };
-  } catch (error) {
-    return axios.isAxiosError(error) && error.response
-      ? {
-          error: error.response.data.message || "Login failed.",
-          status: error.response.status,
-        }
-      : {
-          error: "An unexpected error occurred.",
-        };
-  }
-};
-
-export const nextlogin = async (values: z.infer<typeof LoginSchema>) => {
-  const validatedFields = LoginSchema.safeParse(values);
-  if (!validatedFields.success) {
-    return {
-      error: "Login Failed. Please check your email and password.",
-    };
-  }
-  const { email, password } = validatedFields.data;
-  const payload = { email, password };
-  try {
-    const response = await axios.post(`${apiUrl}/api/v1/auth/login`, payload);
-
-    return response.data;
-  } catch (error) {
-    return axios.isAxiosError(error) && error.response
-      ? {
-          error: error.response.data.message || "Login failed.",
-          status: error.response.status,
-        }
-      : {
-          error: "An unexpected error occurred.",
-        };
-  }
-};
